fix:client and permission delete issues (#202)
* fix:issues

* feat:modify permission index

* feat:update package version

* chore:fix merge error

* feat:udpate package

* update:modify permission name length

+++ src/Services/Masa.Auth.Service.Admin/Application/Subjects/CommandHandler.cs
@@ -17,9 +17,22 @@ public class CommandHandler
     readonly ThirdPartyUserDomainService _thirdPartyUserDomainService;
     readonly IConfiguration _configuration;
     readonly ILogger<CommandHandler> _logger;
+    readonly IUserContext _userContext;
     readonly IDistributedCacheClient _cache;
 
-    public CommandHandler(IUserRepository userRepository, IStaffRepository staffRepository, IThirdPartyIdpRepository thirdPartyIdpRepository, ILdapIdpRepository ldapIdpRepository, StaffDomainService staffDomainService, ILdapFactory ldapFactory, UserDomainService userDomainService, ThirdPartyUserDomainService thirdPartyUserDomainService, IConfiguration configuration, ILogger<CommandHandler> logger, IDistributedCacheClient cache)
+    public CommandHandler(
+        IUserRepository userRepository,
+        IStaffRepository staffRepository,
+        IThirdPartyIdpRepository thirdPartyIdpRepository,
+        StaffDomainService staffDomainService,
+        ILdapFactory ldapFactory,
+        UserDomainService userDomainService,
+        ThirdPartyUserDomainService thirdPartyUserDomainService,
+        ILdapIdpRepository ldapIdpRepository,
+        IMasaConfiguration masaConfiguration,
+        ILogger<CommandHandler> logger,
+        IDistributedCacheClient cache,
+        IUserContext userContext)
     {
         _userRepository = userRepository;
         _staffRepository = staffRepository;
@@ -29,9 +42,10 @@ public class CommandHandler
         _ldapFactory = ldapFactory;
         _userDomainService = userDomainService;
         _thirdPartyUserDomainService = thirdPartyUserDomainService;
-        _configuration = configuration;
+        _configuration = masaConfiguration.Local;
         _logger = logger;
         _cache = cache;
+        _userContext = userContext;
     }
 
     #region User
@@ -116,6 +130,16 @@ public class CommandHandler
     public async Task RemoveUserAsync(RemoveUserCommand command)
     {
         var user = await CheckUserAsync(command.User.Id);
+
+        if (user.Account == "admin")
+        {
+            throw new UserFriendlyException("超级管理员 无法删除");
+        }
+
+        if (user.Id == _userContext.GetUserId<Guid>())
+        {
+            throw new UserFriendlyException("当前用户不能删除自己");
+        }
 
         await _userRepository.RemoveAsync(user);
         await _userDomainService.RemoveAsync(user.Id);
